implemented makeMove methods, +separated logic GameService-MoveService, refactored code
